backend/requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.11.18; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
asgiref==3.8.1; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
certifi==2025.4.26; python_version >= '3.6'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
django==5.2.1; python_version >= '3.10'
django-environ==0.12.0; python_version >= '3.9' and python_version < '4'
djangorestframework==3.16.0; python_version >= '3.9'
drf-yasg==1.21.10; python_version >= '3.6'
fastembed==0.6.1; python_full_version >= '3.9.0'
filelock==3.18.0; python_version >= '3.9'
flatbuffers==25.2.10
frozenlist==1.6.0; python_version >= '3.9'
fsspec==2025.3.2; python_version >= '3.9'
greenlet==3.2.2; python_version >= '3.9'
groq==0.24.0; python_version >= '3.8'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
huggingface-hub==0.31.1; python_full_version >= '3.8.0'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.10; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
jiter==0.9.0; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema-pydantic==0.6
langchain==0.3.25; python_version >= '3.9'
langchain-community==0.3.23; python_version >= '3.9'
langchain-core==0.3.59; python_version >= '3.9'
langchain-groq==0.3.2; python_version >= '3.9' and python_version < '4.0'
langchain-mcp-adapters==0.0.11; python_version >= '3.10'
langchain-openai==0.3.16; python_version >= '3.9'
langchain-text-splitters==0.3.8; python_version >= '3.9' and python_version < '4.0'
langgraph==0.4.3; python_version >= '3.9'
langgraph-checkpoint==2.0.25; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langgraph-prebuilt==0.1.8; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langgraph-sdk==0.1.66; python_version >= '3.9'
langsmith==0.3.42; python_version >= '3.9'
loguru==0.7.3; python_version >= '3.5' and python_version < '4.0'
marshmallow==3.26.1; python_version >= '3.9'
mcp==1.8.0; python_version >= '3.10'
mcp-use==1.2.8; python_version >= '3.11'
mmh3==5.1.0; python_version >= '3.9'
mpmath==1.3.0
multidict==6.4.3; python_version >= '3.9'
mypy-extensions==1.1.0; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
numpy==2.2.5; python_version >= '3.10'
onnxruntime==1.21.1; python_version >= '3.10'
openai==1.78.0; python_version >= '3.8'
orjson==3.10.18; python_version >= '3.9'
ormsgpack==1.9.1; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pillow==11.2.1; python_version >= '3.9'
propcache==0.3.1; python_version >= '3.9'
protobuf==6.30.2; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
py-rust-stemmers==0.1.5
pydantic==2.11.4; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-settings==2.9.1; python_version >= '3.9'
pyreadline3==3.5.4; python_version >= '3.8'
python-dotenv==1.1.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.40; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
sse-starlette==2.3.4; python_version >= '3.9'
starlette==0.46.2; python_version >= '3.9'
sympy==1.14.0; python_version >= '3.9'
tenacity==9.1.2; python_version >= '3.9'
tiktoken==0.9.0; python_version >= '3.9'
tokenizers==0.21.1; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.13.2; python_version >= '3.8'
typing-inspect==0.9.0
typing-inspection==0.4.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.4.0; python_version >= '3.9'
uvicorn==0.34.2; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
win32-setctime==1.2.0; python_version >= '3.5'
xxhash==3.5.0; python_version >= '3.7'
yarl==1.20.0; python_version >= '3.9'
zstandard==0.23.0; python_version >= '3.8'
